* v1.35r, 1 March 2006       + ImageJ now opens 64-bit (double) FITS images.       + Added a "64-bit Real" image type to File/Import/Raw.       + The "java.io.tmpdir" is used when opening ZIP files.       + Added the getResultLabel(), File.open(), File.openAsString() and         File.close() macro functions.       + The 'flags' variable returned by the getCursorLoc() macro function         is updated when the mouse is released.       + Fixed a bug that sometimes caused incorrect unit conversions in the         Image/Properties dialog.       + Fixed a bug that caused Process/Math/Min (or Max) to sometimes not         work correctly with float images.       + Fixed a bug that caused an exception if the IJ.getDirectory()         method was called from the event dispatch thread.       + Fixed a bug that caused the "Info" image property to not be saved         with ZIP files.       + Fixed a bug that prevented stacks >2GB from being saved as TIFF         files. Note that to work with >2GB stacks you need a 64-bit CPU         (e.g., AMD64), a 64-bit OS (e.g., Fedora x86_64) and a 64-bit JVM         (e.g., J2SE for AMD64).

+++ ij/process/FloatProcessor.java
@@ -358,13 +358,13 @@ public class FloatProcessor extends ImageProcessor {
 						break;
 					case MINIMUM:
 						if (v1<value)
-							v2 = (int)value;
+							v2 = (float)value;
 						else
 							v2 = v1;
 						break;
 					case MAXIMUM:
 						if (v1>value)
-							v2 = (int)value;
+							v2 = (float)value;
 						else
 							v2 = v1;
 						break;
